add BalancerUtils#clusterSnapshot
